+++ common/banner.py
@@ -1,49 +1,34 @@
-import sys
-from common.colors import bannerblue, bannerblue2, W, Y, R, end
+from common.colors import bannerblue, bannerblue2, W, Y, G, end
+from common.branding import APP_NAME, EDITION, MAINTAINER, MAINTAINER_ALIAS, REPO_URL
 
 
 def banner():
-    print("""%s
-                                                      
-                    .:.        .:,                    
-                   xM;           XK.                  
-                  dx'            .lO.                 
-                 do                ,0.                
-             .c.lN'      ,  '.     .k0.:'             
-              xMMk;d;''cOM0kWXl,',locMMX.             
-              .NMK.   :WMMMMMMMx    dMMc              
-               lMMO  lWMMMMMMMMMO. lMMO               
-                cWMxxMMMMMMMMMMMMKlWMk                
-                 .xWMMMMMMMMMMMMMMM0,%s                 
-                   .,OMd,,,;0MMMO,.                   
-             .l0O.%sVXVX%sOX.%sVXVX%s0MO%sVXVX%s.0Kd,             
-            lWMMO0%sVXVX0%sOX.%sVXVX%sl%sVXVX%s.VXNMMO            
-           .MMX;.N0%sVXVX0%s0X.%sVXVXVX0%s.0M:.OMMl           
-          .OXc  ,MMO%sVXVX0%sVX%s .VXVX0%s0MMo  ,0X'          
-          0x.  :XMMMk%sVXVX.%sXO.%sVXVX%sdMMMWo.  :X'         
-         .d  'NMMMMMMk%sVXVX%s..%sVXVX0%s.XMMMMWl  ;c         
-            'NNoMMMMMMx%sVXVXVXVXVX0.%sXMMk0Mc            
-           .NMx OMMMMMMd%sVXVXVX%sl%sVXVX%s.NW.;MMc           
-          :NMMd .NMMMMMMd%sVXVX%sdMd,,,,oc ;MMWx          
-          .0MN,  'XMMMMMMo%sVX%soMMMMMMWl   0MW,          
-           .0.    .xWMMMMM:lMMMMMM0,     kc           
-            ,O.     .:dOKXXXNKOxc.      do            
-             '0c        -VulnX-       ,Ol             
-               ;.                     :.   
-
-    %s# Coded By Anouar Ben Saad -%s @anouarbensaad      
-                                           %s"""
-          %
-          (bannerblue, bannerblue2,
-           W, bannerblue2, W, bannerblue2, W, bannerblue2,
-           W, bannerblue2, W, bannerblue2, W, bannerblue2,
-           W, bannerblue2, W, bannerblue2,
-           W, bannerblue2, W, bannerblue2,
-           W, bannerblue2, W, bannerblue2,
-           W, bannerblue2, W, bannerblue2,
-           W, bannerblue2,
-           W, bannerblue2, W, bannerblue2,
-           W, bannerblue2,
-           W, bannerblue2,
-           W, Y, end
-           ))
+    w = 58
+    print(bannerblue + "+" + "=" * w + "+" + end)
+    print(bannerblue + "|" + end + (G + ("  %s  " % APP_NAME).center(w)) + bannerblue + "|" + end)
+    print(bannerblue + "|" + end + (Y + ("  %s  " % EDITION).center(w)) + bannerblue + "|" + end)
+    print(bannerblue + "|" + end + W + "  CMS - Exploits Scan - Dorks - DNS - Ports  ".center(w) + bannerblue + "|" + end)
+    print(bannerblue + "+" + "=" * w + "+" + end)
+    print(bannerblue + "                    .:.        .:," + end)
+    print(bannerblue + "                   xM;           XK." + end)
+    print(bannerblue + "                  dx'            .lO." + end)
+    print(bannerblue + "                 do                ,0." + end)
+    print(bannerblue + "             .c.lN'      ,  '.     .k0.:'" + end)
+    print(bannerblue + "              xMMk;d;''cOM0kWXl,',locMMX." + end)
+    print(bannerblue + "               lMMO  lWMMMMMMMMMO. lMMO" + end)
+    print(bannerblue + "                cWMxxMMMMMMMMMMMMKlWMk" + end)
+    print(bannerblue + "                 .xWMMMMMMMMMMMMMMM0," + bannerblue2 + end)
+    print(bannerblue + "                   .,OMd,,,;0MMMO,." + end)
+    print(
+        bannerblue + "             .l0O." + W + "NEXP" + bannerblue2 + "OX."
+        + W + "NEXP" + bannerblue2 + "0MO" + W + "NEXP" + bannerblue + ".0Kd," + end
+    )
+    print(bannerblue + "             '0c      - " + Y + APP_NAME + bannerblue + " -      ,Ol" + end)
+    print(bannerblue + "               ;.                     :." + end)
+    print("")
+    print(W + "  Repo    : " + G + REPO_URL + end)
+    print(
+        W + "  Credits : " + G + MAINTAINER + W + " (" + MAINTAINER_ALIAS + ")"
+        + W + " | fork: Anouar Ben Saad / VulnX" + end
+    )
+    print("")